rename the granular_dynamics to continuum_dynamics
