+++ tests/user_examples/extra_src/shared/general_continuum/continuum_dynamics/continuum_dynamics_complex.cpp
@@ -1,11 +1,10 @@
-#include "granular_dynamics_complex.h"
-#include "granular_dynamics_complex.hpp"
+#include "continuum_dynamics_complex.hpp"
 
 namespace SPH
 {
 //=================================================================================================//
-namespace granular_dynamics
-{
+	namespace continuum_dyannmics
+	{
 
-}
+	}
 } // namespace SPH
