+++ tests/user_examples/extra_src/shared/general_continuum/continuum_dynamics/continuum_dynamics_inner.hpp
@@ -1,0 +1,9 @@
+#pragma once
+#include "continuum_dynamics_inner.h"
+namespace SPH
+{
+	namespace continuum_dyannmics
+	{
+
+	}
+} // namespace SPH
