+++ tests/user_examples/extra_src/shared/general_continuum/continuum_particles.h
@@ -1,0 +1,41 @@
+#ifndef CONTINUUM_PARTICLES_H
+#define CONTINUUM_PARTICLES_H
+#include "base_particles.h"
+#include "base_particles.hpp"
+#include "general_continuum.h"
+#include "particle_generator_lattice.h"
+namespace SPH
+{
+    class GeneralContinuum;
+
+    class ContinuumParticles : public BaseParticles
+    {
+      public:
+        StdLargeVec<Matd> strain_tensor_;
+        StdLargeVec<Matd> strain_tensor_rate_;
+        StdLargeVec<Vecd> acc_shear_;
+
+        StdLargeVec<Matd> stress_tensor_;
+        StdLargeVec<Matd> stress_tensor_rate_;
+
+        StdLargeVec<Matd> shear_stress_;
+        StdLargeVec<Matd> shear_stress_rate_;
+        StdLargeVec<Matd> velocity_gradient_;
+
+        StdLargeVec<Real> von_mises_stress_;
+
+        StdLargeVec<Vecd> pos0_; /**< initial position */
+        StdLargeVec<Vecd> n_;    /**<  current normal direction */
+        StdLargeVec<Vecd> n0_;   /**<  initial normal direction */
+
+        GeneralContinuum &granular_material_;
+
+        ContinuumParticles(SPHBody &sph_body, GeneralContinuum *granular_material);
+        virtual ~ContinuumParticles(){};
+
+        virtual void initializeOtherVariables() override;
+        virtual ContinuumParticles *ThisObjectPtr() override { return this; };
+    };
+} // namespace SPH
+
+#endif // CONTINUUM_PARTICLES_H
