--- tests/user_examples/extra_src/shared/granular/granular_dynamics/granular_dynamics_inner.hpp
@@ -1,9 +1,0 @@
-#pragma once
-#include "granular_dynamics_inner.h"
-namespace SPH
-{
-namespace granular_dynamics
-{
-
-}
-} // namespace SPH
